stick feature flagged stuff in a module
